refactor(formatters): extract shared defaults, deduplicate fumadocs
