+++ packages/formatters/src/defaults.ts
@@ -1,0 +1,82 @@
+/**
+ * Default formatter implementations shared across multiple formatter packages.
+ *
+ * These functions provide the standard/baseline rendering behaviour.
+ * Individual formatters import what they need and override only what differs.
+ *
+ * @packageDocumentation
+ */
+
+import type {
+  CollapsibleOption,
+  FrontMatterOptions,
+  Maybe,
+  MDXString,
+} from "@graphql-markdown/types";
+import { formatMarkdownFrontmatter } from "@graphql-markdown/helpers";
+import {
+  FRONT_MATTER_DELIMITER,
+  MARKDOWN_EOL,
+  MARKDOWN_EOP,
+} from "@graphql-markdown/utils";
+
+/**
+ * Formats a bullet point separator using the HTML `&bull;` entity.
+ * @param text - Optional text to append after the bullet
+ * @returns Formatted bullet string
+ */
+export const formatMDXBullet = (text = ""): MDXString => {
+  return `&bull;&nbsp;${text}` as MDXString;
+};
+
+/**
+ * Formats a collapsible block as an HTML `<details>` element.
+ * @param option - Configuration for open/close label text
+ * @returns Formatted details element string
+ */
+export const formatMDXDetails = ({
+  dataOpen,
+  dataClose,
+}: CollapsibleOption): MDXString => {
+  return `${MARKDOWN_EOP}<details>${MARKDOWN_EOL}<summary>${dataOpen}</summary>${MARKDOWN_EOL}${MARKDOWN_EOL}\r${MARKDOWN_EOL}${MARKDOWN_EOL}<em>${dataClose}</em>${MARKDOWN_EOL}</details>${MARKDOWN_EOP}` as MDXString;
+};
+
+/**
+ * Formats YAML front matter wrapped in `---` delimiters.
+ * @param _props - Front matter options (unused)
+ * @param formatted - Pre-formatted front matter lines
+ * @returns Formatted front matter block, or empty string if no lines provided
+ */
+export const formatMDXFrontmatter = (
+  _props: Maybe<FrontMatterOptions>,
+  formatted: Maybe<string[]>,
+): MDXString => {
+  return formatMarkdownFrontmatter(
+    formatted,
+    FRONT_MATTER_DELIMITER,
+    MARKDOWN_EOL,
+  ) as MDXString;
+};
+
+/**
+ * Formats a named entity as a backtick code span.
+ * @param name - Entity name
+ * @param parentType - Optional parent type name for qualified references
+ * @returns Formatted entity reference string
+ */
+export const formatMDXNameEntity = (
+  name: string,
+  parentType?: Maybe<string>,
+): MDXString => {
+  const parentName = parentType ? `${parentType}.` : "";
+  return `\`${parentName}${name}\`` as MDXString;
+};
+
+/**
+ * Formats a "specified by" link as a standard Markdown link.
+ * @param url - URL to the specification
+ * @returns Formatted specification link string
+ */
+export const formatMDXSpecifiedByLink = (url: string): MDXString => {
+  return `[Specification ⎘](${url})` as MDXString;
+};
